refactor(ui): reorganize telemetry views into flet_ui/views/telemetry
