+++ flet_ui/views/__init__.py
@@ -1,9 +1,8 @@
 """View components for Flet UI."""
 
-from flet_ui.views.auth_view import AuthView
+from flet_ui.views.telemetry import AuthView, UsageAdoptionDashboardView
 from flet_ui.views.home_view import HomeView
 from flet_ui.views.migration_planner_view import MigrationPlannerPlaceholderView
-from flet_ui.views.usage_adoption_dashboard_view import UsageAdoptionDashboardView
 
 __all__ = [
     "HomeView",
