+++ flet_ui/views/telemetry/auth_view.py
