+++ flet_ui/views/telemetry/sections/base_section_view.py
